Move Links class from lib to models
