+++ app/models/concerns/hathi_links.rb
@@ -1,6 +1,6 @@
 module HathiLinks
   def hathi_links
-    @hathi_links ||= SearchWorks::Links.new([ht_link].compact)
+    @hathi_links ||= Links.new([ht_link].compact)
   end
 
   private
@@ -8,7 +8,7 @@ module HathiLinks
   def ht_link
     return unless ht_links.present?
 
-    SearchWorks::Links::Link.new(
+    Links::Link.new(
       link_text: 'Full text via HathiTrust',
       href: ht_links.url,
       fulltext: true,
