+++ spec/models/concerns/eds_links_spec.rb
@@ -8,11 +8,11 @@ RSpec.describe EdsLinks do
   end
 
   context '#eds_links' do
-    it '#all returns an array of SearchWorks::Links:link' do
+    it '#all returns an array of Links:link' do
       expect(document.eds_links.all).to be_present
 
       document.eds_links.all.each do |link|
-        expect(link).to be_a SearchWorks::Links::Link
+        expect(link).to be_a Links::Link
       end
     end
